+++ docs/changelog.md
@@ -29,6 +29,7 @@ Ends August 31, 2020
 - update handleIconClick in ArrowIcon component so that it's unable to return false.
 - update PlayArea to populate a fixed number of Cards based on the cardOrder prop.
 - remove snapshot testing for App component.
+- add StateMock for PlayArea component testing.
 
 ## Sprint 2
 Ended August 17, 2020